chore(backend split): fix command backend issues
