+++ backend/command/src/app.module.ts
@@ -1,9 +1,8 @@
 import { TypegooseModule } from '@m8a/nestjs-typegoose';
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { CommandsModule } from './commands.module/commands.module';
 import { EventStoreModule } from './eventstore/eventstore.module';
-import { QueriesModule } from './queries.module/queries.module';
+import { ItemsModule } from './items/items.module';
 import { SharedModule } from './shared/shared.module';
 
 @Module({
@@ -29,8 +28,7 @@ import { SharedModule } from './shared/shared.module';
       inject: [ConfigService],
     }),
     EventStoreModule,
-    CommandsModule,
-    QueriesModule,
+    ItemsModule,
     SharedModule,
   ],
 })
