+++ backend/command/src/items/items.module.ts
@@ -1,15 +1,15 @@
 import { TypegooseModule } from '@m8a/nestjs-typegoose';
 import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
-import { ItemsByTag } from '../models/items-by-tag.model';
 import { EventStoreModule } from '../eventstore/eventstore.module';
 import { Item } from '../models/item.model';
+import { ItemsByTag } from '../models/items-by-tag.model';
 import { Profile } from '../models/profile.model';
 import { Suggestion } from '../models/suggestion.model';
 import { Tag } from '../models/tag.model';
 import { CommandHandlers } from './commands';
-import { CommandsController } from './commands.controller';
 import { EventHandlers } from './events';
+import { ItemsController } from './items.controller';
 import { Sagas } from './sagas';
 @Module({
   imports: [
@@ -17,7 +17,7 @@ import { Sagas } from './sagas';
     TypegooseModule.forFeature([Item, ItemsByTag, Tag, Suggestion, Profile]),
     EventStoreModule,
   ],
-  controllers: [CommandsController],
+  controllers: [ItemsController],
   providers: [...CommandHandlers, ...EventHandlers, ...Sagas],
 })
-export class CommandsModule {}
+export class ItemsModule {}
